# Add @interruptable macro