Works like a charm! Scoping was indeed a bit delicate...

In many cases, we could actually skip the extra iteration since everything is recomputed one last time, but there's a argument for that if needed (default to 1 currently).