@pokutta 2025-02-19
very cool - and even a bit better than we discussed with the number of final iterations adjustable. i think 1 is good for some algorithms. stupid question: upon ctrl-c is the "remainder" of the loop iteration still correctly executed? I assume yes?