@pokutta 2025-02-19
> Unfortunately no! The macro only has access to the body of the loop as a whole, so the breaking point is unknown. This may indeed lead to unexpected behaviours, but this is probably fine for testing, which is anyway the purpose of this feature.
> 
> Btw, I've just done a quick performance test and could not see any measurable drawback.

let's discuss this soon - maybe there is a way to ensure that always "full" loops are executed.