@pokutta 2025-02-19
from the logic it seems that the full loop is completed as you only change the variable interrupted via the exception