@SetaR0202 2025-03-17
It seems that `@interruptable` macro is slowing done the code by adding allocations.

I've tested locally the `@interruptable` macro on the `159-add-ctrl-c` branch (**with macro**) and on the `master` branch (**without macro**)  with `@btime` from `BenchmarkTools.jl` over some examples from the `examples` folder.

* `docs_00_fw_visualized.jl` 
  - **with macro**: 453.887 ms (340386 allocations: 18.85 MiB)
  - **without macro**: 344.428 ms (274277 allocations: 14.67 MiB)
* `docs_04_rational_opt.jl` 
  -   **with macro**: 1.129 s (1304290 allocations: 61.97 MiB)
  -  **without macro**: 751.028 ms (1180819 allocations: 55.56 MiB)
* `docs_06_spectrahedron.jl` 
  -   **with macro**: 4.272 s (459838 allocations: 329.59 MiB)
  -  **without macro**: 4.010 s (364488 allocations: 324.34 MiB)
* `docs_07_shifted_norm_polytopes.jl` 
  -   **with macro**: 651.116 ms (272223 allocations: 14.69 MiB)
  -  **without macro**: 369.010 ms (201782 allocations: 10.51 MiB)
* `docs_08_callback_and_tracking.jl` 
  -   **with macro**: 720.030 ms (423697 allocations: 23.35 MiB)
  -  **without macro**: 429.394 ms (315637 allocations: 16.73 MiB)
* `docs_09_extra_vertex_storage.jl`
  -   **with macro**: 3.162 s (1786604 allocations: 99.26 MiB)
  -  **without macro**: 1.777 s (1107891 allocations: 61.71 MiB)
* `docs_10_alternating_methods.jl`
  -   **with macro**: 1.737 s (4307555 allocations: 264.63 MiB)
  -  **without macro**: 1.548 s (4094604 allocations: 252.96 MiB)
* `docs_11_block_coordinate_fw.jl`
  -   **with macro**: 5.188 s (15495988 allocations: 2.80 GiB)
  -  **without macro**: 4.065 s (15049039 allocations: 2.77 GiB)